# [FIXED] Use of `*` and `>` in subjects as literals