@derekcollison 2017-08-16
Does this have a noticeable impact on that functions performance? Should we track separators and token length to make the checks more straightforward?