@kozlovic 2017-08-16
Will look into make it faster. But yes, there is an impact as it is now.

But fundamentally, do you agree that `matchLiteral("foo,*,>", "foo,*,>")`
should return `true` because both strings are literals? (current behavior
is that second param of the function will be interpreting the `*` and `>`
as wildcards and therefore return `false`).