@derekcollison 2017-08-17
Yes I agree, but want to be sensitive to performance impacts.

On Wed, Aug 16, 2017 at 2:16 PM, Ivan Kozlovic <notifications@github.com>
wrote:

> Will look into make it faster. But yes, there is an impact as it is now.
>
> But fundamentally, do you agree that matchLiteral("foo,*,>", "foo,*,>")
> should return true because both strings are literals? (current behavior
> is that second param of the function will be interpreting the * and >
> as wildcards and therefore return false).
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/nats-io/gnatsd/pull/561#issuecomment-322901802>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AAFf8cRCnJWu8ZO1Gt77uZNQskE6CVhkks5sY1wkgaJpZM4O5T8K>
> .
>