@kozlovic 2017-08-17
Agreed.. working on optimizations, but difficult to be as fast as with old code since it was not checking that `*` or `>` were single tokens. Still making some improvements and adding a benchmark test. Will update the PR later.