@derekcollison 2017-08-17
I think the trick may be in tracking "." separators and keeping separate
count on token length, if when you see next ".", or end of subject, etc.
then and only then to you check for pwc or fwc.

On Wed, Aug 16, 2017 at 5:14 PM, Ivan Kozlovic <notifications@github.com>
wrote:

> Agreed.. working on optimizations, but difficult to be as fast as with old
> code since it was not checking that * or > were single tokens. Still
> making some improvements and adding a benchmark test. Will update the PR
> later.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/nats-io/gnatsd/pull/561#issuecomment-322933831>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AAFf8T95srlBa4NGY8bKzDy7Z8jN1Lb9ks5sY4X6gaJpZM4O5T8K>
> .
>