@kozlovic 2017-08-17
I think I tried that but it was slow, will try again. Anytime we check ahead (even if we know that it is in bound, assembly code shows that there is a bound check with possible invocation of panicindex()).