@kozlovic 2017-08-17
I am going to push a change that includes a small optimization, simplification of 'if' statements and a benchmark test. Will continue trying to optimize. Let's not merge for now.