@kozlovic 2017-08-24
Impact on normal bench is not visible. This is a comparison between master and the branch in 1 run only: (note negative value means branch (new) is faster)
```
Ivans-MacBook-Pro:gnatsd ivan$ benchcmp master.txt branch.txt 
benchmark                         old ns/op     new ns/op     delta
Benchmark_____Pub0b_Payload-8     86.8          86.2          -0.69%
Benchmark_____Pub8b_Payload-8     87.2          87.0          -0.23%
Benchmark____Pub32b_Payload-8     95.1          96.3          +1.26%
Benchmark___Pub128B_Payload-8     122           116           -4.92%
Benchmark___Pub256B_Payload-8     180           152           -15.56%
Benchmark_____Pub1K_Payload-8     239           227           -5.02%
Benchmark_____Pub4K_Payload-8     1131          1127          -0.35%
Benchmark_____Pub8K_Payload-8     2292          2314          +0.96%
Benchmark_AuthPub0b_Payload-8     179           175           -2.23%
Benchmark____________PubSub-8     178           177           -0.56%
Benchmark____PubSubTwoConns-8     178           178           +0.00%
Benchmark____PubTwoQueueSub-8     201           200           -0.50%
Benchmark___PubFourQueueSub-8     200           204           +2.00%
Benchmark__PubEightQueueSub-8     201           199           -1.00%
Benchmark___RoutedPubSub_0b-8     496           486           -2.02%
Benchmark___RoutedPubSub_1K-8     1044          1096          +4.98%
Benchmark_RoutedPubSub_100K-8     77286         78541         +1.62%

benchmark                         old MB/s     new MB/s     speedup
Benchmark_____Pub0b_Payload-8     126.75       127.58       1.01x
Benchmark_____Pub8b_Payload-8     218.00       218.39       1.00x
Benchmark____Pub32b_Payload-8     462.85       456.83       0.99x
Benchmark___Pub128B_Payload-8     1154.74      1212.86      1.05x
Benchmark___Pub256B_Payload-8     1488.72      1766.16      1.19x
Benchmark_____Pub1K_Payload-8     4335.65      4558.20      1.05x
Benchmark_____Pub4K_Payload-8     3631.12      3646.54      1.00x
Benchmark_____Pub8K_Payload-8     3579.30      3545.01      0.99x
Benchmark_AuthPub0b_Payload-8     61.18        62.71        1.03x
```
I ran the test another time and the `Benchmark___Pub256B_Payload-8` run is also around 15% faster, not sure why.

@derekcollison